+++ src/ui/log_tab.rs
@@ -442,7 +442,7 @@ impl Component for LogTab<'_> {
 
                 f.render_widget(&*log_revset_textarea, popup_chunks[0]);
 
-                let help = Paragraph::new(vec!["Ctrl+s: save | Escape: cancel".into()])
+                let help = Paragraph::new(vec![self.revset_keybinds.hint("accept").into()])
                     .fg(Color::DarkGray)
                     .alignment(Alignment::Center)
                     .block(